requirements.txt
absl-py==0.12.0
antlr4-python3-runtime==4.8
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
Cython==0.29.22
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.3
entrypoints==0.3
fasteners==0.16
flatbuffers==1.12
future==0.18.2
gast==0.3.3
glfw==2.1.0
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
gym==0.18.3
h5py==2.10.0
hydra-core==1.3.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
importlib-metadata==3.10.1
importlib-resources==5.1.2
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
lockfile==0.12.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.19.5
numpy-stl==2.16.0
oauthlib==3.1.0
olefile==0.46
omegaconf==2.1.0.dev24
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
progressbar2==3.53.1
prometheus-client==0.11.0
prompt-toolkit==3.0.18
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pygame==2.0.1
pyglet==1.5.0
Pygments==2.9.0
PyOpenGL==3.1.5
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-utils==2.5.6
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.5.4
Send2Trash==1.5.0
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
tensorflow==2.4.2
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.10.0
testpath==0.5.0
torch==1.8.1
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.1
zmq==0.0.0
